seperation into 3 classes, added TODO's
